refactor: rename utils/misc.ts to language.ts
